destination-redshift: simplify test class hierarchy
